Structure Rework
rework some of the structure and update macroquad to avoid errors in the example!
