+++ src/cell/mod.rs
@@ -9,4 +9,3 @@ pub struct Cell {
     pub text: Option<String>,
     pub text_color: Option<macroquad::color::Color>
 }
-
